+++ src/Transposition.h
@@ -87,7 +87,7 @@ namespace Transposition {
         u64      _cluster_mask;
         u08      _generation;
 
-        void alloc_aligned_memory (size_t mem_size, size_t alignment);
+        void alloc_aligned_memory (u64 mem_size, u32 alignment);
 
         // free_aligned_memory() free the allocated memory
         void free_aligned_memory ()
@@ -164,7 +164,7 @@ namespace Transposition {
         {
             if (ClearHash && _clusters != NULL)
             {
-                memset (_clusters, 0x00, size_t(_cluster_count * ClusterSize));
+                memset (_clusters, 0x00, _cluster_count * ClusterSize);
                 _generation = 0;
                 sync_cout << "info string Hash cleared." << sync_endl;
             }
@@ -207,11 +207,11 @@ namespace Transposition {
             return u16(full_cluster * 1000 / scan_cluster);
         }
 
-        u32 resize (size_t mem_size_mb, bool force = false);
+        u32 resize (u64 mem_size_mb, bool force = false);
 
         inline u32 resize () { return resize (size (), true); }
 
-        u32 auto_size (size_t mem_size_mb, bool force = false);
+        u32 auto_size (u64 mem_size_mb, bool force = false);
 
         // store() writes a new entry in the transposition table.
         void store (Key key, Move move, Depth depth, Bound bound, Value value, Value eval);
